+++ source/models/AudioSendersModel.hpp
@@ -13,7 +13,6 @@
 #include "ravennakit/core/audio/audio_buffer.hpp"
 #include "ravennakit/ravenna/ravenna_node.hpp"
 #include "util/AudioResampler.hpp"
-#include "util/DriftFilter.hpp"
 #include "util/MessageThreadExecutor.hpp"
 
 #include <juce_audio_devices/juce_audio_devices.h>